Crate peel

Dynamic parsing within trees 🌲 🌳 🌴

Target of this library is to provide a flexible approach in parsing data. This will mainly be done within arena based parser trees which can be modified during runtime. Every parser is using the nom framework for the actual parsing work. A complete source code example can be found within the src/example directory of the crate.

Modules

error
Basic error handling mechanisms

example
Example parsers

parser
General parser descriptions and traits

prelude
Provides sensible imports at all

Structs

Peel
The main peeling structure

PeelResult
General return type of the Peel traversals